feat: add GELU, v_norm, embed_scale, norm_topk_prob for Gemma4/Qwen3 correctness
- Add `gelu()` (via exact erf formula) and `erf` unary op to mlx-sys
- Gemma4 FFN and expert gates now use GELU activation (GEGLU) instead of SiLU (SwiGLU)
- V-norm (no-scale RMSNorm): applied to V before caching for Gemma4 non-KV-shared layers,
  controlled by LayerConfig::v_norm_no_scale from attention_v_norm_no_scale_layers manifest
- Embedding scale: multiply hidden states by sqrt(hidden_size) for Gemma4 (hidden_states_scale)
- Qwen3 MoE norm_topk_prob: renormalise selected expert weights to sum to 1 when enabled
- Fix pre-existing clippy: unnecessary f32 casts in linear_attention.rs, is_multiple_of

+++ crates/ax-engine-mlx/src/linear_attention.rs
@@ -131,8 +131,8 @@ pub fn normalize_linear_attention_qk(
     let inv_scale = (cfg.key_head_dim as f32).powf(-0.5);
     let q_normed = rms_norm(q, None, 1e-6, None);
     let k_normed = rms_norm(k, None, 1e-6, None);
-    let q_scale = scalar_f32_as((inv_scale * inv_scale) as f32, q.dtype());
-    let k_scale = scalar_f32_as(inv_scale as f32, k.dtype());
+    let q_scale = scalar_f32_as(inv_scale * inv_scale, q.dtype());
+    let k_scale = scalar_f32_as(inv_scale, k.dtype());
     (
         multiply(&q_normed, &q_scale, None),
         multiply(&k_normed, &k_scale, None),
